+++ src/lib.rs
@@ -10,7 +10,11 @@ compile_error!(
 
 
 
-use std::{borrow::Cow, ffi::OsStr};
+use std::{
+	borrow::Cow,
+	ffi::OsString,
+	path::{Path, PathBuf},
+};
 
 
 
@@ -47,9 +51,11 @@ pub trait IntoStableEncoding {
 	fn into_stable_encoding(self) -> StableOsString;
 }
 
-/// Converts an `OsString` from an encoding that is stable across rust compiler versions, bypassing data copies if possible
+/// Crates an `OsString` from an encoding that is stable across rust compiler versions, bypassing data copies if possible
 pub trait FromStableEncoding {
 	/// Converts an `OsString` from an encoding that is stable across rust compiler versions, bypassing data copies if possible
+	/// 
+	/// This takes `Into<Cow<[EncodingWidth]>>` so that either a slice can be passed (which always allocates and copies data) or a vec can be passed (which might be able to skip allocating and copying data)
 	///
 	/// # Safety
 	///
@@ -59,11 +65,39 @@ pub trait FromStableEncoding {
 
 
 
-impl<'a> IntoStableEncoding for Cow<'a, OsStr> {
+impl ToStableEncoding for Path {
+	fn to_stable_encoding(&self) -> StableOsString {
+		self.as_os_str().to_stable_encoding()
+	}
+}
+
+impl ToStableEncoding for PathBuf {
+	fn to_stable_encoding(&self) -> StableOsString {
+		self.as_os_str().to_stable_encoding()
+	}
+}
+
+impl IntoStableEncoding for PathBuf {
+	fn into_stable_encoding(self) -> StableOsString {
+		self.into_os_string().to_stable_encoding()
+	}
+}
+
+impl FromStableEncoding for PathBuf {
+	unsafe fn from_stable_encoding<'a>(encoded: impl Into<Cow<'a, [EncodingWidth]>>) -> Self {
+		unsafe { PathBuf::from(OsString::from_stable_encoding(encoded)) }
+	}
+}
+
+impl<'a, T> IntoStableEncoding for Cow<'a, T>
+where
+	T: ToStableEncoding + ToOwned,
+	<T as ToOwned>::Owned: IntoStableEncoding,
+{
 	fn into_stable_encoding(self) -> StableOsString {
 		match self {
-			Cow::Borrowed(os_str) => os_str.to_stable_encoding(),
-			Cow::Owned(os_string) => os_string.into_stable_encoding(),
+			Cow::Borrowed(v) => v.to_stable_encoding(),
+			Cow::Owned(v) => v.into_stable_encoding(),
 		}
 	}
 }
@@ -73,7 +107,7 @@ impl<'a> IntoStableEncoding for Cow<'a, OsStr> {
 #[cfg(test)]
 mod test {
 	use crate::{FromStableEncoding, IntoStableEncoding, ToStableEncoding};
-	use std::ffi::OsString;
+	use std::{ffi::OsString, path::PathBuf};
 
 	#[test]
 	fn basics() {
@@ -86,6 +120,23 @@ mod test {
 
 		let as_os_string_1 = unsafe { OsString::from_stable_encoding(as_stable_1) };
 		let as_os_string_2 = unsafe { OsString::from_stable_encoding(as_stable_2) };
+		assert_eq!(as_os_string_1, as_os_string_2);
+
+		let as_str = as_os_string_1.to_str();
+		assert_eq!(as_str, Some("test"));
+	}
+
+	#[test]
+	fn path_buf() {
+		let start = PathBuf::from(OsString::from("test"));
+		let as_stable_1 = start.to_stable_encoding();
+		let as_stable_2 = start.into_stable_encoding();
+		assert_eq!(as_stable_1, as_stable_2);
+
+		let as_stable_1 = &*as_stable_1; // make sure &[EncodingWidth] can be given to from_stable_encoding()
+
+		let as_os_string_1 = unsafe { PathBuf::from_stable_encoding(as_stable_1) };
+		let as_os_string_2 = unsafe { PathBuf::from_stable_encoding(as_stable_2) };
 		assert_eq!(as_os_string_1, as_os_string_2);
 
 		let as_str = as_os_string_1.to_str();